+++ src/com/company/Rabbit.java
@@ -7,7 +7,7 @@ public class Rabbit extends Animal { // Rabbit inherent from Animal
     private static boolean eatGrain = false;
     private static boolean eatMeat = false;
     private static boolean eatVegetable = true;
-    private int health;
+    private double health;
     public int price = 70;
 
     public Rabbit(String name, String gender){
